MCP_Based_agent/requirements.txt
mcp[cli]
python-dotenv
google-genai
langchain
langchain-mcp-adapters
langchain-google-genai
google-generativeai
requests
